CDRIVER-5745 Enable and address aggressive -Wformat warnings (#1757)
* Format mongoc-stat.c
* Use <PROJECT_NAME>_SOURCE_DIR
* Move source file lists into respective scopes
* cmake: remove redundant include of MongoC-Warnings
* cmake: refactor MongoC-Warnings to set options as a variable
* cmake: reduce scope of mongoc warnings to our source code
* CDRIVER-5745 promote -Wformat warnings to unconditional hard error
* Address -Wformat warnings
* Address portability issues with the integral type of __LINE__
* Address -Wconversion warnings
* Address PRIu8 compatibility issues with GCC in release builds

+++ src/libmongoc/src/mongoc/mongoc-client.c
@@ -72,6 +72,8 @@
 
 #include <mcd-string.h>
 
+#include <inttypes.h>
+
 #undef MONGOC_LOG_DOMAIN
 #define MONGOC_LOG_DOMAIN "client"
 
@@ -784,8 +786,11 @@ mongoc_client_connect (bool buffered,
       base_stream = mongoc_client_connect_unix (host, error);
       break;
    default:
-      bson_set_error (
-         error, MONGOC_ERROR_STREAM, MONGOC_ERROR_STREAM_INVALID_TYPE, "Invalid address family: 0x%02x", host->family);
+      bson_set_error (error,
+                      MONGOC_ERROR_STREAM,
+                      MONGOC_ERROR_STREAM_INVALID_TYPE,
+                      "Invalid address family: 0x%02x",
+                      (unsigned int) host->family);
       break;
    }
